Cleaned up SmartSim folder from SDL_2021 and added new discussion of this approach
