--- 05_Simulation_ML/HandsOn_SmartSim.md
@@ -1,20 +1,0 @@
-# Description
-
-The purpose of this exercise is to show how SmartSim can be used to perform online learning of a neural net model cuncurrently with the simulation. Some slides introducing SmartSim and discussing the software infrastructure of this example are [here](ML_Fortran_SmartSim/SDL21_onlineLearningWithSSIM.pdf). The slides also contain motivation for online learning in the field of turbulence closure modeling as an example where this approach is valuable.
-
-The basic idea is that there is a data producer, a data consumer, a data transporter and an in-memory distributed database.
-As the name suggests, the data producer is an application that produces the training data. In a realistic scenario, this would be a PDE solver integrating the governing equations and generating new solution states to train from as the simulation progresses.
-The data consumer can be any application that consumes (i.e., analyzes, visualizes, compresses or trains from) the data generated by the producer.
-Finally, the data transporter and database are provided by SmartSim ([github page](https://github.com/CrayLabs/SmartSim)). Specifically, the infrastructure library provides API to start, stop and monitor HPC applications from Python and to deplos the distributed in-memory database.
-Additionally, data transport is performed by the SmartRedis client library, which provides clients that can connect to the database from Fortran, C, C++ and Python and an API that enables data transfer to/from the database, commands to act on the data stored in database (e.g., delete), and invoking JIT-traced Python and ML runtimes.
-That fact that the SmartRedis client API is offered in C, C++, Fortran and Python is a significant advantage of this approach because it can be applied to any data producers and consumers written in those languages.
-
-
-For this tutorial, a simple example is used to demonstrate the infrastructure.
-The data producer is a simple Fortran program which iteratively produces training data for a simple polynimial function and sends that data to the database. The fact that new training data is produced iteratively emulates the integration of the equations in a PDE solver.
-Similarly, the data consumer is a a data parallel ML algorithm that trains a simple neural net for the polynomial function. 
-This problem can be run quickly on just a handful of nodes on ThetaKNL, however the implementation of the infrastructure closely resembles a realistic application.
-
-When the example is run successfully, results similar to the ones shown below should be obtained, showing that the neural net model was trained reasonably well.
-
-![predictions](ML_Fortran_SmartSim/fig.png)
