--- 05_Simulation_ML/ML_Fortran_SmartSim/example/src/doConfig.sh
@@ -1,13 +1,0 @@
-#!/bin/bash
-
-CC=cc
-CXX=CC
-FC=ftn
-
-cmake \
--DCMAKE_Fortran_FLAGS="-craype-verbose -g" \
--DCMAKE_CXX_FLAGS="-craype-verbose -g" \
--DCMAKE_C_FLAGS="-craype-verbose -g" \
-./
-
-make
