--- 05_Simulation_ML/ML_Fortran_SmartSim/example/src/env_Theta.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-module swap PrgEnv-intel PrgEnv-gnu
-module load cmake
-export CRAYPE_LINK_TYPE=dynamic
