+++ 05_Simulation_ML/README.md
@@ -6,7 +6,7 @@ Today, we're learning about coupling simulations and machine learning.
 
 1. First, [here](Background.md) we will talk about why we would want to couple simulations & ML and describe ways to do it.
 2. Then, [here](ML_PythonC++_Embedding/README.md) we will work through our hands-on example with one way to couple Python and C++. (written by Romit Maulik)
-3. Finally, we will learn about SmartSim and how to use it [here](HandsOn_SmartSim.md). 
+3. Finally, we will learn about SmartSim and how to use it [here](SmartSim/README.md). 
 
 Bonus:
 For extra tips for using the Python and Numpy C APIs, check out [this page](Tips.md).
