+++ 05_Simulation_ML/SmartSim/example/cleanDir.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+rm orchestrator_0.* smartsim_db.dat 
+rm load_data.* train_model.* *.pt inference.* fig.pdf
+rm 55*
