+++ 05_Simulation_ML/SmartSim/example/src/env_Theta.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+module swap PrgEnv-intel PrgEnv-gnu
+module load cmake
+export CRAYPE_LINK_TYPE=dynamic
